Allow Photino applications to open links in external browser
This commit provides a workaround for issue #111. Users were not able to
open links in an external browser window, only inside the WebView itself.
We now intercept web requests that use the application's scheme (e.g., `app://`)
and, if it does not contain `0.0.0.0` or `localhost`, then we open the link
in the browser.

Resolves tryphotino/photino.Blazor#111